chore: add embed function for bundled cdn dispatch
